requirements.txt
annotated-doc==0.0.2
annotated-types==0.7.0
anyio==4.11.0
blis==1.3.0
cachetools==6.2.1
catalogue==2.0.10
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
cloudpathlib==0.23.0
confection==0.1.5
cymem==2.0.11
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fastapi==0.120.0
google-ai-generativelanguage==0.6.15
google-api-core==2.27.0
google-api-python-client==2.185.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.71.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httplib2==0.31.0
idna==3.11
Jinja2==3.1.6
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
murmurhash==1.0.13
numpy==2.3.4
packaging==25.0
pillow==12.0.0
preshed==3.0.10
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pyparsing==3.2.5
python-dotenv==1.1.1
python-multipart==0.0.20
reportlab==4.4.4
requests==2.32.5
rich==14.2.0
rsa==4.9.1
setuptools==80.9.0
shellingham==1.5.4
smart_open==7.4.1
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.48.0
thinc==8.3.6
tqdm==4.67.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
wasabi==1.1.3
weasel==0.4.1
wrapt==2.0.0
